+++ lib/searchkick/query.rb
@@ -2,6 +2,8 @@ module Searchkick
   class Query
     extend Forwardable
 
+    @@metric_aggs = [:avg, :cardinality, :max, :min, :sum]
+
     attr_reader :klass, :term, :options
     attr_accessor :body
 
@@ -14,7 +16,7 @@ module Searchkick
 
     def initialize(klass, term = "*", **options)
       unknown_keywords = options.keys - [:aggs, :body, :body_options, :boost,
-        :boost_by, :boost_by_distance, :boost_where, :conversions, :debug, :emoji, :exclude, :execute, :explain,
+        :boost_by, :boost_by_distance, :boost_where, :conversions, :conversions_term, :debug, :emoji, :exclude, :execute, :explain,
         :fields, :highlight, :includes, :index_name, :indices_boost, :limit, :load,
         :match, :misspellings, :offset, :operator, :order, :padding, :page, :per_page, :profile,
         :request_params, :routing, :select, :similar, :smart_aggs, :suggest, :slop, :track, :type, :where]
@@ -226,13 +228,15 @@ module Searchkick
         if options[:similar]
           payload = {
             more_like_this: {
-              fields: fields,
               like_text: term,
               min_doc_freq: 1,
               min_term_freq: 1,
               analyzer: "searchkick_search2"
             }
           }
+          if fields != ["_all"]
+            payload[:more_like_this][:fields] = fields
+          end
         elsif all
           payload = {
             match_all: {}
@@ -280,7 +284,13 @@ module Searchkick
 
             match_type =
               if field.end_with?(".phrase")
-                field = field.sub(/\.phrase\z/, ".analyzed")
+                field =
+                  if field == "_all.phrase"
+                    "_all"
+                  else
+                    field.sub(/\.phrase\z/, ".analyzed")
+                  end
+
                 :match_phrase
               else
                 :match
@@ -289,18 +299,25 @@ module Searchkick
             shared_options[:operator] = operator if match_type == :match
             shared_options[:slop] = slop if match_type == :match_phrase
 
+            exclude_analyzer = nil
+            exclude_field = field
+
             if field == "_all" || field.end_with?(".analyzed")
               shared_options[:cutoff_frequency] = 0.001 unless operator == "and" || misspellings == false
               qs.concat [
                 shared_options.merge(analyzer: "searchkick_search"),
                 shared_options.merge(analyzer: "searchkick_search2")
               ]
+              exclude_analyzer = "searchkick_search2"
             elsif field.end_with?(".exact")
               f = field.split(".")[0..-2].join(".")
               queries_to_add << {match: {f => shared_options.merge(analyzer: "keyword")}}
+              exclude_field = f
+              exclude_analyzer = "keyword"
             else
               analyzer = field =~ /\.word_(start|middle|end)\z/ ? "searchkick_word_search" : "searchkick_autocomplete_search"
               qs << shared_options.merge(analyzer: analyzer)
+              exclude_analyzer = analyzer
             end
 
             if misspellings != false && match_type == :match
@@ -327,10 +344,13 @@ module Searchkick
 
             if options[:exclude]
               must_not =
-                options[:exclude].map do |phrase|
+                Array(options[:exclude]).map do |phrase|
                   {
                     match_phrase: {
-                      field => phrase
+                      exclude_field => {
+                        query: phrase,
+                        analyzer: exclude_analyzer
+                      }
                     }
                   }
                 end
@@ -367,7 +387,7 @@ module Searchkick
                       boost_mode: "replace",
                       query: {
                         match: {
-                          "#{conversions_field}.query" => term
+                          "#{conversions_field}.query" => options[:conversions_term] || term
                         }
                       }
                     }.merge(script_score)
@@ -433,7 +453,7 @@ module Searchkick
         set_aggregations(payload) if options[:aggs]
 
         # suggestions
-        set_suggestions(payload) if options[:suggest]
+        set_suggestions(payload, options[:suggest]) if options[:suggest]
 
         # highlight
         set_highlights(payload, fields) if options[:highlight]
@@ -476,17 +496,22 @@ module Searchkick
     def set_fields
       boost_fields = {}
       fields = options[:fields] || searchkick_options[:searchable]
+      default_match = options[:match] || searchkick_options[:match] || :word
       fields =
         if fields
           fields.map do |value|
-            k, v = value.is_a?(Hash) ? value.to_a.first : [value, options[:match] || searchkick_options[:match] || :word]
+            k, v = value.is_a?(Hash) ? value.to_a.first : [value, default_match]
             k2, boost = k.to_s.split("^", 2)
             field = "#{k2}.#{v == :word ? 'analyzed' : v}"
             boost_fields[field] = boost.to_f if boost
             field
           end
-        else
+        elsif default_match == :word
           ["_all"]
+        elsif default_match == :phrase
+          ["_all.phrase"]
+        else
+          raise ArgumentError, "Must specify fields"
         end
       [boost_fields, fields]
     end
@@ -556,12 +581,18 @@ module Searchkick
       payload[:indices_boost] = indices_boost
     end
 
-    def set_suggestions(payload)
-      suggest_fields = (searchkick_options[:suggest] || []).map(&:to_s)
+    def set_suggestions(payload, suggest)
+      suggest_fields = nil
 
-      # intersection
-      if options[:fields]
-        suggest_fields &= options[:fields].map { |v| (v.is_a?(Hash) ? v.keys.first : v).to_s.split("^", 2).first }
+      if suggest.is_a?(Array)
+        suggest_fields = suggest
+      else
+        suggest_fields = (searchkick_options[:suggest] || []).map(&:to_s)
+
+        # intersection
+        if options[:fields]
+          suggest_fields &= options[:fields].map { |v| (v.is_a?(Hash) ? v.keys.first : v).to_s.split("^", 2).first }
+        end
       end
 
       if suggest_fields.any?
@@ -573,6 +604,8 @@ module Searchkick
             }
           }
         end
+      else
+        raise ArgumentError, "Must pass fields to suggest option"
       end
     end
 
@@ -637,6 +670,12 @@ module Searchkick
             date_histogram: {
               field: histogram[:field],
               interval: interval
+            }
+          }
+        elsif metric = @@metric_aggs.find { |k| agg_options.has_key?(k) }
+          payload[:aggs][field] = {
+            metric => {
+              field: agg_options[metric][:field] || field
             }
           }
         else